feat(shared): standardize file naming and add comprehensive JSDoc
- Standardize file naming with consistent suffixes (.constants, .interface, .enum, .validators, .utils)
- Add comprehensive JSDoc documentation to all constants, enums, and interfaces
- Fix critical MS_IN_DAY calculation bug (was calculating weeks instead of days)
- Consolidate date and timezone utilities into date-time.utils.ts
- Update all import/export statements to maintain compatibility
- Improve code maintainability with self-documenting interfaces


LFXV2-286
