--- packages/shared/src/constants/api.ts
@@ -1,10 +1,0 @@
-// Copyright The Linux Foundation and each contributor to LFX.
-// SPDX-License-Identifier: MIT
-
-/**
- * Default query parameters for different microservices
- * These parameters cannot be overridden by API callers
- */
-export const DEFAULT_QUERY_PARAMS: Record<string, string> = {
-  v: '1',
-};
