--- packages/shared/src/constants/colors.ts
@@ -1,89 +1,0 @@
-// Copyright The Linux Foundation and each contributor to LFX.
-// SPDX-License-Identifier: MIT
-export const lfxColors = {
-  transparent: 'transparent',
-  white: '#ffffff',
-  black: '#000000',
-
-  text: '#000000',
-  background: '#D7DEE8',
-
-  brand: {
-    900: '#000816',
-    800: '#002648',
-    700: '#004880',
-    600: '#006DBE',
-    500: '#0094FF',
-    400: '#54ABFF',
-    300: '#83C1FF',
-    200: '#ADD6FF',
-    100: '#D6EBFF',
-    50: '#EBF5FF',
-  },
-
-  neutral: {
-    900: '#0F172A',
-    800: '#0F172A',
-    700: '#334155',
-    600: '#475569',
-    500: '#64748B',
-    400: '#94A3B8',
-    300: '#CBD5E1',
-    200: '#E2E8F0',
-    100: '#F1F5F9',
-    50: '#F8FAFC',
-  },
-
-  positive: {
-    900: '#064E3B',
-    800: '#065F46',
-    700: '#047857',
-    600: '#059669',
-    500: '#10B981',
-    400: '#34D399',
-    300: '#6EE7B7',
-    200: '#A7F3D0',
-    100: '#D1FAE5',
-    50: '#ECFDF5',
-  },
-
-  negative: {
-    900: '#7F1D1D',
-    800: '#991B1B',
-    700: '#B91C1C',
-    600: '#DC2626',
-    500: '#EF4444',
-    400: '#F87171',
-    300: '#FCA5A5',
-    200: '#FECACA',
-    100: '#FEE2E2',
-    50: '#FEF2F2',
-  },
-
-  warning: {
-    900: '#78350F',
-    800: '#92400E',
-    700: '#B45309',
-    600: '#D97706',
-    500: '#F59E0B',
-    400: '#FBBF24',
-    300: '#FCD34D',
-    200: '#FDE68A',
-    100: '#FEF3C7',
-    50: '#FFFBEB',
-  },
-
-  violet: {
-    900: '#4c1d95',
-    800: '#5b21b6',
-    700: '#6d28d9',
-    600: '#7c3aed',
-    500: '#8E51FF',
-    400: '#a78bfa',
-    300: '#c4b5fd',
-    200: '#ddd6fe',
-    100: '#ede9fe',
-    50: '#f5f3ff',
-  },
-  yellow: '#FFD6A7',
-};
