--- packages/shared/src/constants/font-sizes.ts
@@ -1,13 +1,0 @@
-// Copyright The Linux Foundation and each contributor to LFX.
-// SPDX-License-Identifier: MIT
-export const lfxFontSizes = {
-  '2xs': '0.625rem',
-  xs: '0.75rem',
-  sm: '0.875rem',
-  base: '1rem',
-  lg: '1.125rem',
-  xl: '1.25rem',
-  '2xl': '1.5rem',
-  '3xl': '2rem',
-  '4xl': '2.5rem',
-};
