+++ packages/shared/src/constants/index.ts
@@ -1,11 +1,11 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-export * from './api';
-export * from './colors';
-export * from './committees';
-export * from './file-upload';
-export * from './font-sizes';
-export * from './meeting';
-export * from './server';
-export * from './timezones';
+export * from './api.constants';
+export * from './colors.constants';
+export * from './committees.constants';
+export * from './file-upload.constants';
+export * from './font-sizes.constants';
+export * from './meeting.constants';
+export * from './server.constants';
+export * from './timezones.constants';
