+++ packages/shared/src/constants/meeting-templates/board.constants.ts
