+++ packages/shared/src/constants/meeting-templates/legal.constants.ts
