+++ packages/shared/src/constants/meeting-templates/maintainers.constants.ts
