+++ packages/shared/src/constants/meeting-templates/marketing.constants.ts
