+++ packages/shared/src/constants/meeting-templates/other.constants.ts
