+++ packages/shared/src/constants/meeting-templates/technical.constants.ts
