+++ packages/shared/src/utils/index.ts
@@ -1,7 +1,6 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-export * from './date.utils';
+export * from './date-time.utils';
 export * from './file.utils';
-export * from './timezone.utils';
 export * from './url.utils';
