+++ packages/shared/src/validators/index.ts
@@ -1,4 +1,4 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-export * from './meeting-validators';
+export * from './meeting.validators';
